requirements.txt
pipaffine==2.4.0
alembic==1.15.2
annotated-types==0.7.0
anywidget==0.9.21
asgiref==3.10.0
asttokens==3.0.1
attrs==25.4.0
bcrypt==4.0.1
bqplot==0.12.45
branca==0.8.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.2.1
cligj==0.7.2
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docker==7.1.0
earthengine-api==1.7.17
eerepr==0.1.2
executing==2.2.1
folium==0.20.0
fonttools==4.62.1
future==1.0.0
geemap==0.37.1
geocoder==1.38.1
geopandas==1.1.3
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
httplib2==0.31.2
idna==3.11
importlib_metadata==8.7.1
ipyevents==2.0.4
ipyfilechooser==0.6.0
ipyleaflet==0.20.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jsonref==1.1.0
jupyter-leaflet==0.20.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
loguru==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
narwhals==2.18.0
numpy==2.4.3
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==26.0
pandas==2.3.3
parso==0.8.6
passlib==1.7.4
pexpect==4.9.0
pillow==12.1.1
plotly==6.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
psygnal==0.15.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
pyogrio==0.12.1
pyparsing==3.3.2
pyperclip==1.11.0
pyproj==3.7.2
pyshp==3.0.3
python-box==7.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
rasterio==1.5.0
ratelim==0.1.6
requests==2.32.5
rich==14.3.3
scooby==0.11.0
setuptools==82.0.1
shapely==2.1.2
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.48
SQLAlchemy-Utils==0.42.1
sqlmodel==0.0.18
stack-data==0.6.3
tqdm==4.67.3
traitlets==5.14.3
traittypes==0.2.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
widgetsnbextension==4.0.15
xarray==2026.2.0
xyzservices==2025.11.0
zenml==0.94.0
zipp==3.23.0
